+++ src/OpenSage.Mods.Bfme/Gui/AptSkirmish.cs
@@ -1,14 +1,11 @@
-﻿using OpenSage.Content;
-using OpenSage.Gui.Apt;
+﻿using OpenSage.Gui.Apt;
 using OpenSage.Gui.Apt.ActionScript;
 
 namespace OpenSage.Mods.Bfme.Gui
 {
-    [AptCallbacks(SageGame.Bfme)]
-    public class AptSkirmish
+    [AptCallbacks(SageGame.Bfme, SageGame.Bfme2, SageGame.Bfme2Rotwk)]
+    static class AptSkirmish
     {
-        public MapCache CurrentMap { get; private set; }
-
         // Called after the initialization has been performed
         public static void OnInitialized(string param, ActionContext context, AptWindow window, Game game)
         {
