+++ src/OpenSage.Mods.Bfme/Gui/MpGameSetup.cs
@@ -1,9 +1,9 @@
 ﻿using OpenSage.Gui.Apt;
 using OpenSage.Gui.Apt.ActionScript;
 
-namespace OpenSage.Mods.Bfme.Gui
+namespace OpenSage.Mods.Bfme2.Gui
 {
-    [AptCallbacks(SageGame.Bfme)]
+    [AptCallbacks(SageGame.Bfme, SageGame.Bfme2, SageGame.Bfme2Rotwk)]
     static class MpGameSetup
     {
         // Called after the initialization has been performed
