+++ src/OpenSage.Mods.Bfme/Gui/PalantirCommandUI.cs
@@ -1,9 +1,9 @@
 ﻿using OpenSage.Gui.Apt;
 using OpenSage.Gui.Apt.ActionScript;
 
-namespace OpenSage.Mods.Bfme.Gui
+namespace OpenSage.Mods.Bfme2.Gui
 {
-    [AptCallbacks(SageGame.Bfme)]
+    [AptCallbacks(SageGame.Bfme, SageGame.Bfme2, SageGame.Bfme2Rotwk)]
     class PalantirCommandUI
     {
         private static NLog.Logger logger = NLog.LogManager.GetCurrentClassLogger();
