+++ src/OpenSage.Mods.Bfme2/Bfme2RotwkDefinition.cs
@@ -5,10 +5,9 @@ using OpenSage.Gui;
 using OpenSage.Gui.Apt;
 using OpenSage.Gui.ControlBar;
 using OpenSage.Gui.CommandListOverlay;
-using OpenSage.Mods.Bfme2;
 using System.IO;
 
-namespace OpenSage.Mods.Bfme2Rotwk
+namespace OpenSage.Mods.Bfme2
 {
     public class Bfme2RotwkDefinition : IGameDefinition
     {
@@ -29,7 +28,7 @@ namespace OpenSage.Mods.Bfme2Rotwk
             new RegistryKeyPath(@"SOFTWARE\Electronic Arts\Electronic Arts\The Battle for Middle-earth II", "Language")
         };
 
-        public string Identifier { get; } = "bfme2";
+        public string Identifier { get; } = "bfme2_rotwk";
 
         public IMainMenuSource MainMenu { get; } = new AptMainMenuSource("MainMenu.apt");
         public IControlBarSource? ControlBar { get; }
