--- src/OpenSage.Mods.Bfme2/Gui/AptSkirmish.cs
@@ -1,25 +1,0 @@
-﻿using OpenSage.Gui.Apt;
-using OpenSage.Gui.Apt.ActionScript;
-
-namespace OpenSage.Mods.Bfme2.Gui
-{
-    [AptCallbacks(SageGame.Bfme2)]
-    static class AptSkirmish
-    {
-        // Called after the initialization has been performed
-        public static void OnInitialized(string param, ActionContext context, AptWindow window, Game game)
-        {
-        }
-
-        public static void Exit(string param, ActionContext context, AptWindow window, Game game)
-        {
-            var aptWindow = game.LoadAptWindow("MainMenu.apt");
-            game.Scene2D.AptWindowManager.QueryTransition(aptWindow);
-        }
-
-        public static void DisableComponents(string param, ActionContext context, AptWindow window, Game game)
-        {
-            // do we need to hide the buttons from MainMenu here?
-        }
-    }
-}
