--- src/OpenSage.Mods.Bfme2Rotwk/Gui/MpGameSetup.cs
@@ -1,14 +1,0 @@
-﻿using OpenSage.Gui.Apt;
-using OpenSage.Gui.Apt.ActionScript;
-
-namespace OpenSage.Mods.Bfme2Rotwk.Gui
-{
-    [AptCallbacks(SageGame.Bfme2Rotwk)]
-    static class MpGameSetup
-    {
-        // Called after the initialization has been performed
-        public static void OnReadyPress(string param, ActionContext context, AptWindow window, Game game)
-        {
-        }
-    }
-}
