--- src/OpenSage.Mods.Bfme2Rotwk/OpenSage.Mods.Bfme2Rotwk.csproj
@@ -1,6 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <ItemGroup>
-    <ProjectReference Include="..\OpenSage.Game\OpenSage.Game.csproj" />
-    <ProjectReference Include="..\OpenSage.Mods.Bfme2\OpenSage.Mods.Bfme2.csproj" />
-  </ItemGroup>
-</Project>
