+++ src/OpenSage.Mods.BuiltIn/GameDefinition.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using OpenSage.Mods.Generals;
 using OpenSage.Mods.Bfme;
 using OpenSage.Mods.Bfme2;
-using OpenSage.Mods.Bfme2Rotwk;
 
 namespace OpenSage.Mods.BuiltIn
 {
