+++ src/OpenSage.Mods.Generals/Gui/GameOptionsUtil.cs
@@ -1,10 +1,11 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using NLog;
 using OpenSage.Content;
 using OpenSage.Content.Translation;
+using OpenSage.Gui;
 using OpenSage.Gui.Wnd;
 using OpenSage.Gui.Wnd.Controls;
 using OpenSage.Logic;
