+++ src/OpenSage.sln
@@ -52,8 +52,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenSage.Rendering", "OpenS
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenSage.Mathematics.Tests", "OpenSage.Mathematics.Tests\OpenSage.Mathematics.Tests.csproj", "{DCD60D8E-36FF-4C5C-B71B-19EA41A5FDF3}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OpenSage.Mods.Bfme2Rotwk", "OpenSage.Mods.Bfme2Rotwk\OpenSage.Mods.Bfme2Rotwk.csproj", "{81FA71B5-6D48-457D-A85D-E0B0B7827245}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenSage.IO", "OpenSage.IO\OpenSage.IO.csproj", "{CE6B09A7-CE42-4F9F-9274-B16979DC61A1}"
 EndProject
 Global
@@ -142,10 +140,6 @@ Global
 		{DCD60D8E-36FF-4C5C-B71B-19EA41A5FDF3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DCD60D8E-36FF-4C5C-B71B-19EA41A5FDF3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DCD60D8E-36FF-4C5C-B71B-19EA41A5FDF3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{81FA71B5-6D48-457D-A85D-E0B0B7827245}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{81FA71B5-6D48-457D-A85D-E0B0B7827245}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{81FA71B5-6D48-457D-A85D-E0B0B7827245}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{81FA71B5-6D48-457D-A85D-E0B0B7827245}.Release|Any CPU.Build.0 = Release|Any CPU
 		{CE6B09A7-CE42-4F9F-9274-B16979DC61A1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{CE6B09A7-CE42-4F9F-9274-B16979DC61A1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CE6B09A7-CE42-4F9F-9274-B16979DC61A1}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -163,7 +157,6 @@ Global
 		{8483402D-4DB2-4E7B-AEAA-9091A3F0255C} = {FFC49638-8FFE-4C6E-B2B2-AC7D942498D6}
 		{09B2C509-4033-433F-9442-4D81B930C715} = {DA1B5AB5-4CC9-49CB-ACF1-34CF6A69D5BE}
 		{5D829E83-DB6D-4C5B-9692-7F69C518CB16} = {DA1B5AB5-4CC9-49CB-ACF1-34CF6A69D5BE}
-		{81FA71B5-6D48-457D-A85D-E0B0B7827245} = {9410748F-ABB1-45AF-98E3-98858A8C0010}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {DBD83F1E-3F41-45D9-AACC-62E00826AD5C}
